+++ R/inputs.R
@@ -77,7 +77,7 @@ clearableTextInput <- function(
 #'         textInputGroup("id2", "right", right_text = "b"),
 #'         textInputGroup("id3", "both", left_text = "$", right_text = ".00"),
 #'         textInputGroup("id4", "none"),
-#'         textInputGroup("id5", "icon", left_text = icon("home")),
+#'         textInputGroup("id5", "icon", left_text = icon("house")),
 #'     )
 #'
 #'     server <- function(input, output, session) {
@@ -156,7 +156,7 @@ textInputGroup <-function(
 #'       textButton(
 #'         textId = "tbtn-icon",
 #'         label = "change icon and color",
-#'         btn_icon = icon("home"),
+#'         btn_icon = icon("house"),
 #'         class = "btn-warning" # pass to the button
 #'       ),
 #'       textButton(