+++ R/spsUIcollections.R
@@ -801,7 +801,7 @@ heightMatcher <- function(div1, div2, isID=TRUE){
 #'         h1("Scroll the page..."),
 #'         lapply(1: 100, function(x) br()),
 #'         spsGoTop("default"),
-#'         spsGoTop("mid", right = "50%",  bottom= "50%", icon = icon("home"), color = "red"),
+#'         spsGoTop("mid", right = "50%",  bottom= "50%", icon = icon("house"), color = "red"),
 #'         spsGoTop("up", right = "95%",  bottom= "95%", icon = icon("arrow-up"), color = "green")
 #'     )
 #'