+++ examples/demo/ui.R
@@ -12,8 +12,8 @@ sidebar <- dashboardSidebar(
     menuItem("Input Control", tabName = "text", icon = icon("font")),
     menuItem("Image Display", tabName = "images", icon = icon("images")),
     menuItem("Buttons", tabName = "buttons", icon = icon("square")),
-    menuItem("Progress control", tabName = "progress", icon = icon("tasks")),
-    menuItem("Upload components", tabName = "upload_file", icon = icon("file-upload")),
+    menuItem("Progress control", tabName = "progress", icon = icon("bars-progress")),
+    menuItem("Upload components", tabName = "upload_file", icon = icon("file-arrow-up")),
     menuItem("Server Components", tabName = "server_col", icon = animateIcon("server"), badgeLabel = "0.4 upd"),
     menuItem("Misc", tabName = "other", icon = icon("border-all"), badgeLabel = "0.3 Updates")
   )